Update to main Python script
Made several updates to the main script to improve UX and functionality.
Fixed an error where the file extensions would be cut off and omitted when the file name exceeded 42 characters.
Improved the files renaming convention.
Added a file counter to the progress bar to allow real-time tracking of the files being processed.